[GITHUB-40] Add support for custom environment varaibles
When for instance you want to use NewRelic java agent, you can use

```YAML
  - role: sansible.kafka
    sansible_kafka_environment_vars:
      - "NEWRELIC_OPTS=\"-javaagent:/home/{{ sansible_kafka_user }}/newrelic/newrelic.jar\""
      - "export KAFKA_OPTS=\"${KAFKA_OPTS} ${NEWRELIC_OPTS}\""
```
