+++ README.md
@@ -5,7 +5,6 @@ Develop: [![Build Status](https://travis-ci.org/sansible/kafka.svg?branch=develo
 
 * [Installation and Dependencies](#installation-and-dependencies)
 * [Tags](#tags)
-* [Maintenance scripts](#maintenance-scripts)
 * [Examples](#examples)
 
 This roles installs Apache Kafka server.
@@ -32,38 +31,11 @@ and run `ansible-galaxy install -p ./roles -r roles.yml`
 
 ### AWS Setup
 
-This role has AWS support built in, it supports two methods for
-deployment/discovery.
-
-#### AWS Cluster Autodiscovery
-
-This method is designed for use with a single ASG controlling a cluster of
-Kafka instances, the idea being that instances can come and go without issue.
-
-The [AWS Autodiscover script](/files/aws_cluster_autodiscover) allows machines
-to pick an ID and hostname/Route53 entry from a predefined list, AWS tags are
-used to mark machines that have claimed an ID/host.
-
-This script allows for a static set of hostnames with consistent IDs to be
-maintained across a dynamic set of instances in an ASG.
-
-```YAML
-- role: sansible.kafka
-  sansible_kafka_aws_cluster_autodiscover_enabled: yes
-  sansible_kafka_aws_cluster_autodiscover_hosts:
-    - 01.kafka.io.internal
-    - 02.kafka.io.internal
-    - 03.kafka.io.internal
-  sansible_kafka_aws_cluster_autodiscover_lookup_filter: "Name=tag:Environment,Values=dev Name=tag:Role,Values=kafka"
-  sansible_kafka_aws_cluster_autodiscover_r53_zone_id: xxxxxxxx
-  # A ZK cluster behind an ELB
-  sansible_kafka_zookeeper_hosts:
-    - zookeeper.app.internal
-```
+This role has AWS support built in for deployment/discovery.
 
 #### AWS Tag Discovery
 
-Designed for instances that are stacially defined either as direct EC2
+Designed for instances that are statically defined either as direct EC2
 instances or via a single ASG per instance.
 
 The broker.id is derived from a tag attached to the instance, you can turn on
@@ -87,44 +59,6 @@ This role uses two tags: **build** and **configure**
 * `configure` - Configure and ensures that the Kafka service is running.
 
 
-## Maintenance scripts
-
-These scripts are used in conjunction with the
-[AWS Cluster Autodiscovery](aws-cluster-autodiscovery) deployment method.
-
-* kafka_maintenance_at_start
-
-  Intention behind this script is to introduce a new node to the cluster and
-  evenly redistribute data. It's included in Configure stage of Ansible role.
-  The new node contacts Zookeeper (ZK) and requests all brokers IDs currently
-  holding data.  Once information is received json file is generated and
-  information provided to ZK.
-
-
-* kafka_maintenance_at_stop
-
-  Intention behind this script is to allow node to remove itself from cluster
-  during shutdown and evenly redistribute data to remaining nodes. Script is
-  triggered by stop_kafka included in relevant runlevels.
-  Node contacts Zookeeper (ZK) and requests all brokers IDs currently holding
-  data.  Once information is received json file is generated and information
-  provided to ZK.
-
-* remove_dns_record
-
-  After kafka_maintenance_at_stop is executed during shutdown (stop_kafka) node
-  removes itself from Route53 (AWS).
-
-* TODO:
-  Becaue kafka_maintenance_start/stop are almost identical they can be merged.
-  Depends on use argument could be provided.
-  Example:
-  kafka_maintenance at_start
-
-  To remove node from Route53 (AWS) Ansible module can be also used.
-  This will require tests.
-
-
 ## Examples
 
 ```YAML
@@ -142,8 +76,28 @@ These scripts are used in conjunction with the
 
   roles:
     - name: sansible.kafka
-      sansible_kafka_aws_cluster_autodiscover_hosts:
+      sansible_kafka_zookeeper_hosts:
         - my.zookeeper.host
+```
+
+```YAML
+- name: Install Kafka with NewRelic integration
+  hosts: sandbox
+
+  pre_tasks:
+    - name: Update apt
+      become: yes
+      apt:
+        cache_valid_time: 1800
+        update_cache: yes
+      tags:
+        - build
+
+  roles:
+    - name: sansible.kafka
+      sansible_kafka_environment_vars:
+        - "NEWRELIC_OPTS=\"-javaagent:/home/{{ sansible_kafka_user }}/newrelic/newrelic.jar\""
+        - "export KAFKA_OPTS=\"${KAFKA_OPTS} ${NEWRELIC_OPTS}\""
 ```
 
 If you just want to test Kafka service build both Zookeeper and Kafka on the
