+++ defaults/main.yml
@@ -4,6 +4,7 @@ sansible_kafka_apache_mirror: https://archive.apache.org/dist/
 sansible_kafka_aws_cluster_assigned_id_enabled: no
 sansible_kafka_aws_cluster_assigned_id_tag_name: instanceindex
 sansible_kafka_conf_dir: /home/kafka/etc
+sansible_kafka_environment_vars: []
 sansible_kafka_group: kafka
 sansible_kafka_heap_opts: "-Xmx{{ (ansible_memtotal_mb / 2) | int }}m -Xms{{ (ansible_memtotal_mb / 2) | int }}m"
 sansible_kafka_interface_advertise: ~
@@ -18,7 +19,7 @@ sansible_kafka_port: 9092
 sansible_kafka_server_properties: {}
 sansible_kafka_tarball_location: /home/kafka/tmp
 sansible_kafka_user: kafka
-sansible_kafka_version_kafka: 2.0.0
+sansible_kafka_version_kafka: 2.1.0
 sansible_kafka_version_scala: 2.11
 sansible_kafka_wait_for_kafka_port: 120
 sansible_kafka_zookeeper_connection_timeout_ms: 1000000
