+++ molecule/default/playbook.yml
@@ -4,6 +4,10 @@
   hosts: all
 
   # Take a look at hosts vars in molecule.yml as well
+  vars:
+    sansible_kafka_environment_vars:
+      - "NEWRELIC_OPTS=\"-javaagent:/home/{{ sansible_kafka_user }}/newrelic/newrelic.jar\""
+      - "export TEST_KAFKA_OPTS=\"${KAFKA_OPTS} ${NEWRELIC_OPTS}\""
 
   roles:
     - role: sansible.zookeeper
