+++ templates/environment.j2
@@ -6,3 +6,5 @@ GROUP=$NAME
 KAFKA_HEAP_OPTS="{{ sansible_kafka_heap_opts }}"
 JMX_PORT="{{ sansible_kafka_jmx_port }}"
 MAX_OPEN_FILES="{{ sansible_kafka_max_open_files }}"
+
+{{ sansible_kafka_environment_vars | join("\n") }}
